wrap link around entire button

+++ frontend/components/course/syllabus/lesson-feed.tsx
@@ -1,14 +1,14 @@
 "use client";
 import React from "react";
 
-import {z } from "zod"
+import { z } from "zod"
 import { zodResolver } from "@hookform/resolvers/zod"
-import {useEffect} from "react";
-import { AuthContext} from "@/providers/AuthProvider";
+import { useEffect } from "react";
+import { AuthContext } from "@/providers/AuthProvider";
 import useUserConvexData from "@/hooks/useUserConvexData";
-import {useRouter} from "next/navigation";
-import {api} from "@/convex/_generated/api";
-import { useQuery,useMutation } from "convex/react";
+import { useRouter } from "next/navigation";
+import { api } from "@/convex/_generated/api";
+import { useQuery, useMutation } from "convex/react";
 import { useForm } from "react-hook-form";
 
 import { Id } from "@/convex/_generated/dataModel";
@@ -33,8 +33,8 @@ import {
     AlertDialogHeader,
     AlertDialogTitle,
     AlertDialogTrigger,
-  } from "@/components/ui/alert-dialog"
-  
+} from "@/components/ui/alert-dialog"
+
 import { toast } from "@/components/ui/use-toast";
 
 
@@ -45,7 +45,7 @@ interface Props {
 
 
 export default function LessonFeed(
-    {courseID, courseName}: Props
+    { courseID, courseName }: Props
 ) {
 
     const courseLessons = useQuery(api.functions.lessons.getLessonsByCourseID, {
@@ -74,49 +74,49 @@ export default function LessonFeed(
                 })
             }
         });
-        
+
         //Toast
         toast({
             title: "Lesson Deleted",
             description: `Lesson has been deleted`,
             variant: "default"
         });
-        
+
 
     };
 
     return (
         <>
             <ScrollArea className="h-[750px] w-full rounded-md border ">
-                
-                    <CardHeader>
-                        <CardTitle>
-                            Syllabus
-                        </CardTitle>
-                        <CardDescription>
-                            Lessons for {courseName} 
-                        </CardDescription>
-                        <div className="flex flex-row gap-2">
+
+                <CardHeader>
+                    <CardTitle>
+                        Syllabus
+                    </CardTitle>
+                    <CardDescription>
+                        Lessons for {courseName}
+                    </CardDescription>
+                    <div className="flex flex-row gap-2">
+                        <Link href={`/courses/${courseID}/syllabus`}>
                             <Button >
-                                <Link href={`/courses/${courseID}/syllabus`}>
-                                    Add Lesson
-                                </Link>
+                                Add Lesson
                             </Button>
-                            <Button variant="outline">
-                                    {/* Make Functional Delete Button */}
-                                    Re-arrange
-                            </Button>
-                            <Button variant="outline">
-                                    {/* Make Functional Delete Button */}
-                                    Print
-                            </Button>
-                        </div>
-                    </CardHeader>
-                    <div className="flex flex-col gap-4 pb-16">
-                   
+                        </Link>
+                        <Button variant="outline">
+                            {/* Make Functional Delete Button */}
+                            Re-arrange
+                        </Button>
+                        <Button variant="outline">
+                            {/* Make Functional Delete Button */}
+                            Print
+                        </Button>
+                    </div>
+                </CardHeader>
+                <div className="flex flex-col gap-4 pb-16">
+
                     {courseLessons?.map((lessons: any) => {
                         return (
-                            <Card 
+                            <Card
                                 key={lessons._id}
                                 className="border ml-3 mr-3 "
                             >
@@ -129,37 +129,37 @@ export default function LessonFeed(
                                     </CardDescription>
                                 </CardHeader>
                                 <CardContent className="flex flex-row gap-3">
-                                    <Button >
-                                        <Link href={`/courses/${courseID}/lessons?select=${lessons._id}`}>
+                                    <Link href={`/courses/${courseID}/lessons?select=${lessons._id}`}>
+                                        <Button >
                                             View
-                                        </Link>
-                                    </Button>
+                                        </Button>
+                                    </Link>
+                                    <Link href={`${window.location.pathname}?edit=${lessons._id}`}>
                                     <Button variant="outline">
-                                        <Link href={`${window.location.pathname}?edit=${lessons._id}`}>
                                             Edit
-                                        </Link>
                                     </Button>
-                                    
+                                    </Link>
+
 
                                     <AlertDialog>
-                                    <AlertDialogTrigger><div className="h-10 py-2 px-4 border border-input bg-background hover:bg-accent hover:text-accent-foreground inline-flex items-center justify-center whitespace-nowrap rounded-md text-sm font-medium ring-offset-background transition-colors focus-visible:outline-none focus-visible:ring-2 focus-visible:ring-ring focus-visible:ring-offset-2 disabled:pointer-events-none disabled:opacity-50 ">Delete</div></AlertDialogTrigger>
-                                    <AlertDialogContent>
-                                        <AlertDialogHeader>
-                                        <AlertDialogTitle>Are you absolutely sure?</AlertDialogTitle>
-                                        <AlertDialogDescription>
-                                            This action cannot be undone. This will permanently delete the lesson
-                                            and all its data from our servers.
-                                        </AlertDialogDescription>
-                                        </AlertDialogHeader>
-                                        <AlertDialogFooter>
-                                        <AlertDialogCancel>Cancel</AlertDialogCancel>
-                                        <AlertDialogAction
-                                        onClick={() => handleDeleteLesson(
-                                            lessons._id,
-                                            lessons.Number
-                                        )}>Continue</AlertDialogAction>
-                                        </AlertDialogFooter>
-                                    </AlertDialogContent>
+                                        <AlertDialogTrigger><div className="h-10 py-2 px-4 border border-input bg-background hover:bg-accent hover:text-accent-foreground inline-flex items-center justify-center whitespace-nowrap rounded-md text-sm font-medium ring-offset-background transition-colors focus-visible:outline-none focus-visible:ring-2 focus-visible:ring-ring focus-visible:ring-offset-2 disabled:pointer-events-none disabled:opacity-50 ">Delete</div></AlertDialogTrigger>
+                                        <AlertDialogContent>
+                                            <AlertDialogHeader>
+                                                <AlertDialogTitle>Are you absolutely sure?</AlertDialogTitle>
+                                                <AlertDialogDescription>
+                                                    This action cannot be undone. This will permanently delete the lesson
+                                                    and all its data from our servers.
+                                                </AlertDialogDescription>
+                                            </AlertDialogHeader>
+                                            <AlertDialogFooter>
+                                                <AlertDialogCancel>Cancel</AlertDialogCancel>
+                                                <AlertDialogAction
+                                                    onClick={() => handleDeleteLesson(
+                                                        lessons._id,
+                                                        lessons.Number
+                                                    )}>Continue</AlertDialogAction>
+                                            </AlertDialogFooter>
+                                        </AlertDialogContent>
                                     </AlertDialog>
 
                                 </CardContent>
@@ -167,13 +167,13 @@ export default function LessonFeed(
                         );
                     })}
 
-                    </div>
-                    
+                </div>
 
 
-            
+
+
             </ScrollArea>
-            
+
         </>
     )
 }
